requirements-dev.txt
-r requirements.txt

# Development dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3
flake8==6.1.0
black==23.9.1
isort==5.12.0
pytest==7.4.2
pytest-django==4.5.2
coverage==7.3.2
Sphinx==7.2.6 